Rename files and refactor project structure, wishlist navigation , back arrows

+++ app/src/main/java/com/example/stylish/ui/components/Homepage/RatingBar.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 
 @Composable
-fun RatingBar(rating: Float) {
+fun RatingBar(rating: Double) {
     Row {
         repeat(5) { index ->
             val starColor = if (index < rating.toInt()) Color(0xFFFFD700) else Color.Gray
